+++ src/protocolgame.h
@@ -162,7 +162,7 @@ class ProtocolGame final : public Protocol
 		void sendFYIBox(const std::string& message);
 
 		void sendDistanceShoot(const Position& from, const Position& to, uint8_t type);
-		void sendMagicEffect(const Position& pos, uint8_t type);
+		void sendMagicEffect(const Position& pos, uint16_t type);
 		void sendCreatureHealth(const Creature* creature);
 		void sendCreatureName(const Creature* creature);
 		void sendSkills();
